eslint properly
