+++ expo-eslint/resources/.prettierrc.json
@@ -1,0 +1,6 @@
+{
+  "tabWidth": 2,
+  "singleQuote": true,
+  "jsxSingleQuote": true,
+  "bracketSameLine": false
+}
